[0.2.4.5] IWidget interface | rework Modifier API

+++ src/enums/Gravity.h
@@ -5,14 +5,16 @@
 
 namespace circuits {
 
-    enum class Gravity : uint32_t {
+    enum class Gravity : uint8_t {
         None        = 0,
+
         Left        = 1 << 0,
         HCenter     = 1 << 1,
         Right       = 1 << 2,
         Top         = 1 << 4,
         VCenter     = 1 << 5,
         Bottom      = 1 << 6,
+
         Center      = HCenter   | VCenter,
         TopLeft     = Top       | Left,
         TopRight    = Top       | Right,
@@ -22,13 +24,13 @@ namespace circuits {
 
     inline Gravity operator|(Gravity a, Gravity b) {
         return static_cast<Gravity>(
-            static_cast<uint32_t>(a) | static_cast<uint32_t>(b)
+            static_cast<uint8_t>(a) | static_cast<uint8_t>(b)
         );
     }
 
     inline Gravity operator&(Gravity a, Gravity b) {
         return static_cast<Gravity>(
-            static_cast<uint32_t>(a) & static_cast<uint32_t>(b)
+            static_cast<uint8_t>(a) & static_cast<uint8_t>(b)
         );
     }
 
@@ -38,8 +40,8 @@ namespace circuits {
     }
 
     inline bool any(Gravity g, Gravity mask) {
-        return (static_cast<uint32_t>(g) &
-                static_cast<uint32_t>(mask)) != 0;
+        return (static_cast<uint8_t>(g) &
+                static_cast<uint8_t>(mask)) != 0;
     }
 
 }
